Remove all traces of users, user_sessions and admin/
As this is a very simple (basically static) site, Jonas Myrlund and I
have decided not to worry about users and such. Changes to the database
will be made in the seed script, and deployed with rake db:seed.

This is a Big and Dangerous Commit, but it seems that everything still
works.