Update the menu
This is in response to RT #58424 (again). I've had to make a slight change
to the Pages controller because there was a problem with the layout; the
number of dishes became greater than the number of dishes per column, so
the last dish was put in a (fourth) column of its own.